Simplify collectd config plugins
Note that I've not included the dimensions property yet since there's some work
to have those values properly formatted in the templates. This does remove
dimension support from write_http because before it was assumed the user would
pre-format the query string themselves but now we need to format it from a map.