Use pod name as the service name
Before the service name was set to the rule name. Templates use the service
name to identify the service (e.g. the redis instance was being given the name
redis-image-default for all monitored instances).

Instead set the service name to something that makes sense for the observer
(pod name in the case of Kubernetes) and leave it untouched.