Implement filesystem polling watcher
Kubernetes config maps are implemented with two symbolic links (a symbolink
link to a symbolink link to the actual file). This makes using a filesystem
watcher either impossible or difficult since watching a symlink watches for
changes to the underyling file but not to the symlink being switched out.

This doesn't yet implement directory watching. That will come in a subsequent
PR.