Implement agent configuration reloading
This implement agent reloading such that if the main configuration file or any
merge configuration files change the agent and its plugins are reloaded.
Plugins can be added, removed, or have their configuration changed without
restarting.

Not all properties in the collectd monitor plugin are currently being reloaded.
That will be fixed in an upcoming change.