Implement unified file/zookeeper configuration loading
This enables configuration to be loaded from any kind of "store". For now this
only includes the filesystem and zookeeper but any store libkv would be trivial
to add now. The store handles the loading of files as well as sending change
notifications.

There are still a number of TODOs to clean up but it seems to basically work.

Also there's more work to be done for plugins to be able to use the new store
infrastructure. This is just for config files.