Implement asset loading for plugins from stores
Adds directory-based asset loading for plugins. Instead of having it managed by
the manager each plugin that needs assets will have to create its own
AssetSyncer. The AssetSyncer is there to keep track of multiple sources and
merges it into one view so the plugin doesn't have to worry about tracking
multiple watches itself.

Also restructures plugin interface a bit. Reload() is gone and replaced with
Configure(). Configure() is called on what used to be reload as well as after
start. This is so there doesn't have to be two different load paths.