Fixing GenericJMX Host config
It will now set the `host` dimension of the collectd datapoint to
`<collectd_hostname>[hostHasService=<service_name>]` where the
`<collectd_hostname>` is gotten from Collectd directly and the `<service_name>`
is specified by a new config option called `ServiceName` to the GenericJMX
plugin.

Also adding an `InstanceSuffix` option to the Connection block that allows
templates to append dimensions to the end of the plugin_instance value.

Updated Cassandra, Kafka, and ActiveMQ templates to use these new features.