Tweaking Plugin Interface
Wrapping plugin instance instead of using embedded types.  This gives more
flexibility in how we create plugins and reduces plugin coupling to the agent.

Removing the `Start` method on plugin interface.  This was unnecessary and
mostly unused.  Any startup work should be done in either the factory function
or Configure.

Simplifying the plugin factory function to be a simple no-arg builder that
returns `interface{}`.  Passing the config into the factory was unnecessary
and confusing since there is a Configure method made for that.