Docker Multi-stage Build
This replaces the two previous Dockerfiles and build script and replaces it
with a single Dockerfile and a much simplified build script.  Should be much
more efficient on changes and hopefully will increase dev cycle efficiency.

Compiling collectd and using only compiled artifacts instead of the previous
technique of installing collectd package in the final image.  The way that the
build script selectively copied collectd plugins was rather convoluted.

Installing collectd Python plugins from a script instead of directly in the
Dockerfile.  There is now a file (`scripts/collectd-plugins.yaml`) that defines
what plugins we want in the neo-agent and how to get them.

Getting rid of the previous startup script that supported adding files through
envvars.  They didn't appear to be used and it would be much better to just
extend this Docker image to include files rather than using those.