Unifying Configuration
Doing all configuration in a single place and getting rid of the merge config.

Getting rid of external configuration of the pipeline -- there is now a clear
distinction between observers that provide services and monitors that consume
services.  Proxying is handled as part of the core neoagent and is no longer a
plugin.

Bringing all collectd plugins up to being full-fledged monitors of their own.
They still use collectd under the covers, but are configured just like any
other monitor will be.  Their templates are generated with `make templates` and
are compiled directly to go source files to be compiled with the rest of the
app.

Rearranging packages so that most of the core neoagent resides in the `core`
package.

Implementing a more unified filtering capacity that will facilitate future
monitors as well as provide filtering for existing monitors.