Simplifying Manual Endpoint Configuration
Removing the whole merging of config from manually defined services with the
top level monitor config.  It was too complicated and offered very little
value.  This now allows much easier validation of monitor config.

Also got rid of the whole conversion from config struct to maps when going to
collectd templates.  This was also too complex for what little value it
offered in terms of adding convenience to the templates.  Things are more
explicit now, which I think is better.