Non-containerized Agent
Rewriting Dockerfile to use scratch image as final image so that the filesystem
can just be dumped out and will be exactly equivalent to what runs in the
containerized version.

Adding make target to do that dump, as well as add in a simple startup script
for the agent that bind mounts in some useful things and starts the agent in a
chroot jail.  We can add more sophisticated init scripts for
systemd/upstart/sysV later that would replace much of the run script.

Also adding in delve in the dev image for easier debugging.

Temporarily excising neopy from the agent core so that we don't have to worry
about libzmq.  I would like to replace zmq with simpler UDP unix sockets when
more of a need for neopy arises.