Tweaking versioning
Got rid of the VERSIONS file since collectd version info can just go in the
Dockerfile and the agent version can be determined form the git tags in the
repo.

Any inter-release commits are considered to be the previous verision with the
"-post" suffix since it is unknown until release whether it will be a patch,
minor, or major increment.

The next step is to get a Jenkins job that automatically does releases on
command.