Watch Docker events in Docker observer instead of polling
To do better service discovery on AWS ECS, we need to get the inspect
output from the Docker API, which requires making an API call per
container.  This is not scalable to do with polling, so now the observer
will use the same event watching logic that the docker-container-stats
monitor uses.  That code was refactored to make it more flexible.

Now ports discovered by the Docker observer will be based on on exposed
ports in the container config instead of ports mapped in the network
settings.  This makes it work a lot better when using non-bridged
network plugins to Docker.