Patch up the bundle bins and libs on build and install
At compile/build time, we set the rpath to the right thing on all
binaries and shared object (so) files.  We can do with using relative
paths ($ORIGIN).  This eliminates the need to set LD_LIBRARY_PATH within
the agent when calling out to the python and collectd subprocesses.  It
also means that subprocesses of these subprocesses will not have
LD_LIBRARY_PATH set by default (e.g. if Python creates a subprocess by
default it inherits the env of the parent process, which is undesirable
in many cases).

At install time, we set the interpreter on the executable binaries to
the bundled ld-linux.so interpreter.  This means that we don't have to
invoke collectd and python directly with the loader, which means that
the process name will be the normal thing, instead of including the
loader in the name.

Also rearranged the bundle contents a bit to hopefully be a bit easier
to find things.