Add all libnss*.so files to bundle (#726)
There was a subtle issue where the bundle's libc was dynamically loading
libnss_compat.so from the host's standard library path at runtime, which
was then chain loading libnss_files.so, which was causing linker errors
due to the bundle's interpreter being used to load it.  Now we include
all libnss libraries so that won't happen.

Also adding two integration tests for the raw bundle to hopefully catch
issues like this in the future.