Filtering (#777)
Make the agent inherently filter monitor outputs so that only a default set of metrics is emitted by default by monitors.  Also standardizes the way monitors should be configured to handle non-default metrics.  This will be backwards compatible with existing deployments since the operation of it is hidden behind a feature flag.