Add cluster property sync to host dimensions at startup (#836)
* Add cluster property sync to host dimensions at startup

If the user specifies the `cluster` config option, the agent will add
this property to the set of properties on the host-id dimensions
identified by the agent.  If there is a platform-specific dimension
(e.g. AWSUniqueId), then the property will only be set on that dimension
and *not* on `host`.

Right now there is no retry on the dimension update so if it fails the
cluster property won't be attempted to be set until the next agent
restart.

* Add retry on existing dimension fetch

* Add ability to skip generating platform host ids