Add signalfx-forwarder monitor
It is an expansion of the trace-forwarder monitor.  Allows datapoints
and trace spans to be sent to a local agent instance and get
buffered/forwarded to where ever the agent is configured to send
datapoints or spans.

There is also the added benefit of getting host id dimensions added to
datapoints and spans that come through the forwarder.

The `trace-forwader` monitor now is just an alias for this monitor but
should continue to work the same.