Add portless endpoints to all observers
These allow for auto-discovery of containers that don't have any
ports associated with them, as well as monitoring ports that are not
exposed or declared in a way that an observer would know about them.

In general, users should not be using discovery rules that don't
contain a `port` variable.  If they were relying on the fact that a
container only had a single port exposed, and thus only one monitor
instance being created, this will break things because now there will be
a new port-less instance that will match a discovery rule without
`port`, resulting in multiple instances of the monitor being created,
which is probably not desired.