Fix azure tag sync when hostname is not lower cased
On the asynctasker side we lower case the resource id when reporting values. So
that the resource ids match up we also need to lower case them on the agent
side.

Also fix an issue where the build would fail removing genmetadata.go when
running from a shared folder (theory is that there's a path too long somewhere
in the directory structure) so just limit the cleanup to internal/monitors.