Switch from govaluate to expr
They are mostly overlapping syntax but expr is more powerful (e.g. map
accessors) and is not a dead project like govaluate.

The main incompatibility that has to be smoothed over is the =~ regexp
match operator.  This is supported by doing a find/replace on that
operator with the expr equivalent matches operator.