RFC: Example of collectable monitor
One of the most common patterns is for a monitor to do collection of metrics every
`intervalSeconds`. This adds an (optional) Collectable interface where if
the monitor implements it we will call it every `intervalSeconds`.

This simplifies monitors that follow this pattern as well as make it easier to implement
other features like checking how long collect intervals are taking and logging if
they're taking longer than `intervalSeconds` which would result in lagging datapoints.