Improve performance of host observer
This skips the uid lookup since we don't use it. See upstream PRs
for some more background:

https://github.com/shirou/gopsutil/issues/784
https://github.com/shirou/gopsutil/pull/783

This requires upgrading gopsutil which changed some structs which in
turn requires upgrading telegraf. Apparently the `Stolen` field was never set
(https://github.com/shirou/gopsutil/pull/677)

Even with this change there are still some performance issues in the host
observer on machines with a large number of connections. Something like this
change would be required to deal with the remaining issues:

https://github.com/shirou/gopsutil/issues/695