Add disk_op.total metric and don't send all
The disk_ops.total metric was generated by the signalfx-metadata monitor and was
not migrated over when the diskio monitor was made GA.

Also don't send all diskio metrics by default -- not sure why this was enabled before.
This will break anybody who was using non-default metrics but they should not have
been on in the first place.