Make time.Duration safer in monitor configs (#1168)
This makes current usage of time.Duration in monitor configs safer as before
if the user specified an integer it'd be interpreted as an nanoseconds which
is unlikely what people want.

Instead of using time.Duration in a monitor config the developer should use
timeutil.Duration. An integer will be interpreted as seconds (instead of
nanoseconds) and if it's a string it'll be parsed via ParseDuration. Also
if it's a string but it is able to be parsed as an integer we also interpret
it as seconds.