Make Helm monitor config easier to manage (#1230)
Add some 4.x/5.x compatibility logic to make the agent configured
properly for both major versions.

This also adds a config option called 'configureStandardMonitoring' that
defaults to 'true' that keeps the standard set of host + docker + K8s
monitors.  If set to 'false', then the only monitors that will be added
are the ones specified in the 'monitors' config option.

Removes all of the 'built-in' monitor configuration from the default value of
'montiors'.

Also preparing the chart for 5.0.0 release.