Add kubelet-summary monitor to replace kubelet-stats
kubelet-stats uses endpoints that are now deprecated in K8s 1.18 and will
be disabled permanently in a future release.

Apparently the /stats/summary endpoint is being left and is what the metrics server in
K8s still relies on so it will probably be safe for a while.

Also updated the Helm chart to conditionally use the new monitor in
place of kubelet-stats if the agent is sufficiently new.