Add Windows Docker Support
This is just a simple Go compilation of the agent for Windows,
without any Python or Java subsystems.

In order to make this work, there need to be a special value `bypass` to
the `-service` cli flag that causes the Windows service framework to be
skipped and the agent executed directly.

The Helm chart is also updated to deploy the Windows agent container
more seamlessly.