Allow k8s-api observer's discovery settings to be independently configured via Helm
When using the Helm chart, it is not currently possible to set the k8s-api observer's discoverAllPods or discoverNodes parameters except by setting the isServerless parameter to true. There are some situations where the observer needs to be configured to discover all pods or nodes in a non-serverless environment, thus this change will set the values to true if the new discoverAllPods/discoverNodes setting is true or if isServerless is set to true.

With the current default values, this does result in a minor change to the ConfigMap - removing discoverAllPods and discoverNodes if they are not set to true. This does not result in a behavioral change, as the observer itself will treat these values as false if not explicitly defined.